+++ requirements.txt
@@ -1,0 +1,9 @@
+fastapi==0.115.7            # FastAPI 서버
+uvicorn==0.34.0             # FastAPI 실행용 ASGI 서버
+openai==1.60.1              # DeepSeek API 연동
+mediapipe==0.10.20          # 손 좌표 추출
+opencv-contrib-python==4.10.0.84  # OpenCV 추가 기능 포함
+python-dotenv==1.0.1        # 환경변수 관리
+gtts==2.3.1                 # Google TTS
+pydantic==2.10.6            # 데이터 모델 검증
+websockets==14.1            # WebSocket 처리
