+++ src/slts/hand_recognition.py
@@ -1,0 +1,1 @@
+# 손 좌표 -> 수어 단어 변환
