+++ src/stsl/speech_to_text.py
@@ -1,0 +1,1 @@
+# 음성 -> 텍스트 변환
