+++ src/stsl/text_to_sign.py
@@ -1,0 +1,1 @@
+# 텍스트 -> 수어 애니메이션 변환
